requirements.txt
cached-conv==2.5.0
click==8.2.1
einops==0.8.1
gin-config==0.5.0
librosa==0.11.0
lmdb==1.6.2
matplotlib==3.10.3
mir_eval==0.8.2
music2latent==0.1.6
nnAudio==0.3.3
nn_tilde==1.6.0
numpy==2.1.2
pandas==2.2.3
pretty_midi==0.2.10
pyloudnorm==0.1.1
pytorch-lightning==2.5.1.post0
resampy==0.4.3
scipy==1.15.3
seaborn==0.13.2
tensorboard==2.19.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0